Features: Fork, RateLimit, Not
Ports features from my working branch which have proven to be the
most valueable. These are core building blocks that are very nice
to have out of the box. Fork is a group tick like Sequence and
Selector, but is inherently stateful, unlike those other two.
While the stateless pattern with optional statefulness via Async
and Sync is very powerful, Fork should be easier to use.