refactor(sequence+selector): pass through errors unmodified
Any error values will not be munged directly by the Sequence or Selector
implementations of Tick. The primary reason for this removal was it's general
uselessness vs the amount of potential log spam that may be generated, when
propagating an error upwards through a large tree.

I also noticed I had flipped the comments somehow, oops.