feat(Memorize): concise stateful sequence selector etc
Memorize is a stateful tick wrapper, as an alternative to the (admittedly,
rather obtuse) Sync implementation, which wraps child nodes directly. Both
implementations address the same core problem case; Memorize is simpler and
more intuitive, and should be used unless the flexibility provided by Sync is
necessary.

Also improves documentation, including a link to the academic paper that this
simplified mechanism was modeled after.